requirements.txt
streamlit
pandas
matplotlib
mariadb
python-dotenv
SQLAlchemy
mariadb-connector-python
